update persistence packages

+++ Koben.IPRestrictor/Controllers/IPRestrictorController.cs
@@ -3,13 +3,9 @@ using Koben.IPRestrictor.Services.IpDataService.Interfaces;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Net;
 using Umbraco.Cms.Web.BackOffice.Controllers;
 using Umbraco.Cms.Web.Common.Attributes;
-using Umbraco.Cms.Web.Common.Controllers;
 
 namespace Koben.IPRestrictor.Controllers
 {
@@ -36,19 +32,32 @@ namespace Koben.IPRestrictor.Controllers
 			{
 				var filtered = data.GroupBy(x => x.Alias).Select(x => x.First());
 
-				var currentIps = _whitelistedIpDataService.GetAll();
+				var currentIps = _whitelistedIpDataService.GetAll()?.ToList();
 
-				var toDelete = currentIps.Where(x => !filtered.Any(y => y.Alias == x.Alias));
-				foreach (var ip in toDelete)
+				var toDelete = currentIps?.Where(x => filtered.All(y => y.Alias != x.Alias)).ToList();
+				if (toDelete?.Any() == true)
 				{
-					_whitelistedIpDataService.Delete(ip.Id);
+					foreach (var ip in toDelete)
+					{
+						_whitelistedIpDataService.Delete(ip.Id);
+					}
+
+					_logger.LogInformation("Deleted the following IP addresses: {0}", string.Join(", ", toDelete.Select(ip => $"{ip.Alias} ({ip.FromIp} > {ip.ToIp})")));
 				}
 
-				var toInsert = filtered.Where(x => IPAddress.TryParse(x.FromIp, out var a) && IPAddress.TryParse(x.ToIp, out var b) && !currentIps.Any(y => y.Alias == x.Alias));
-				var inserted = _whitelistedIpDataService.Insert(toInsert);
+				var toInsert = filtered.Where(x => currentIps != null && IPAddress.TryParse(x.FromIp, out var a) && IPAddress.TryParse(x.ToIp, out var b) && currentIps.All(y => y.Alias != x.Alias)).ToList();
+
+				if (toInsert.Any())
+				{
+					_whitelistedIpDataService.Insert(toInsert);
+
+					_logger.LogInformation("Added the following IP addresses: {0}", string.Join(", ", string.Join(", ", toInsert.Select(ip => $"{ip.Alias} ({ip.FromIp} > {ip.ToIp})"))));
+				}
 			}
-			catch
+			catch (Exception ex)
 			{
+				_logger.LogError(ex, ex.Message);
+
 				return new StatusCodeResult(StatusCodes.Status500InternalServerError);
 			}
 			return Ok();
@@ -60,7 +69,7 @@ namespace Koben.IPRestrictor.Controllers
 			try
 			{
 				var data = _whitelistedIpDataService.GetAll();
-				return Ok(data.Cast<WhiteListedIpDto>());
+				return Ok(data?.Cast<WhiteListedIpDto>());
 			}
 			catch (Exception ex)
 			{
